Update listingDTO to match listing entity

+++ backend/src/main/java/com/propertystack/homematch/Listing/ListingDTO.java
@@ -15,7 +15,7 @@ public class ListingDTO {
     BigDecimal price;
     Integer sqft;
     Integer beds;
-    Double baths;
+    BigDecimal baths;
 
     String listingUrl;
